# Refactor namespace store