--- a/vault/namespaces_store.go
+++ b/vault/namespaces_store.go
@@ -55,7 +55,6 @@ type NamespaceStore struct {
 	// initialization time and persisted throughout the lifetime of the
 	// instance. Entries should not be returned directly but instead be
 	// copied to prevent modification.
-	namespaces           []*NamespaceEntry
 	namespacesByPath     map[string]*NamespaceEntry
 	namespacesByUUID     map[string]*NamespaceEntry
 	namespacesByAccessor map[string]*NamespaceEntry
@@ -153,12 +152,10 @@ func (ns *NamespaceStore) loadNamespacesLocked(ctx context.Context) error {
 	// invalidation this will pre-allocate enough space to reload everything
 	// as we'll likely be essentially in sync already. However, at startup, this
 	// will mostly just give us space for the root namespace.
-	allNamespaces := make([]*NamespaceEntry, 0, len(ns.namespaces)+1)
 	namespacesByPath := make(map[string]*NamespaceEntry, len(ns.namespacesByPath)+1)
 	namespacesByUUID := make(map[string]*NamespaceEntry, len(ns.namespacesByUUID)+1)
 	namespacesByAccessor := make(map[string]*NamespaceEntry, len(ns.namespacesByAccessor)+1)
 	rootNs := &NamespaceEntry{Namespace: namespace.RootNamespace}
-	allNamespaces = append(allNamespaces, rootNs)
 	namespacesByPath[rootNs.Namespace.Path] = rootNs
 	namespacesByUUID[rootNs.UUID] = rootNs
 	namespacesByAccessor[rootNs.Namespace.ID] = rootNs
@@ -183,7 +180,6 @@ func (ns *NamespaceStore) loadNamespacesLocked(ctx context.Context) error {
 				return false, fmt.Errorf("failed to decode namespace %v (page %v / index %v): %w", path, page, index, err)
 			}
 
-			allNamespaces = append(allNamespaces, &namespace)
 			namespacesByPath[namespace.Namespace.Path] = &namespace
 			namespacesByUUID[namespace.UUID] = &namespace
 			namespacesByAccessor[namespace.Namespace.ID] = &namespace
@@ -198,7 +194,6 @@ func (ns *NamespaceStore) loadNamespacesLocked(ctx context.Context) error {
 		return err
 	}
 
-	ns.namespaces = allNamespaces
 	ns.namespacesByPath = namespacesByPath
 	ns.namespacesByUUID = namespacesByUUID
 	ns.namespacesByAccessor = namespacesByAccessor
@@ -250,7 +245,7 @@ func (ns *NamespaceStore) SetNamespace(ctx context.Context, namespace *Namespace
 
 // setNamespaceLocked must be called while holding a write lock over the
 // NamespaceStore. This function unlocks the lock once finished.
-func (ns *NamespaceStore) setNamespaceLocked(ctx context.Context, namespace *NamespaceEntry) error {
+func (ns *NamespaceStore) setNamespaceLocked(ctx context.Context, nsEntry *NamespaceEntry) error {
 	// If we are creating a net-new namespace, we have to unlock before
 	// creating required mounts as the mount type will call
 	// GetNamespaceByAccessor. In that case, we will manually call
@@ -269,12 +264,12 @@ func (ns *NamespaceStore) setNamespaceLocked(ctx context.Context, namespace *Nam
 	}()
 
 	// Copy the entry before validating and potentially mutating it.
-	entry := namespace.Clone()
+	entry := nsEntry.Clone()
 	if err := entry.Validate(); err != nil {
 		return fmt.Errorf("failed validating namespace: %w", err)
 	}
 
-	index := -1
+	var exists bool
 	if entry.UUID == "" {
 		id, err := ns.assignIdentifier(entry.Namespace.Path)
 		if err != nil {
@@ -287,38 +282,36 @@ func (ns *NamespaceStore) setNamespaceLocked(ctx context.Context, namespace *Nam
 			return err
 		}
 	} else {
-		// Ensure we have no conflicts for paths.
-		for idx, existing := range ns.namespaces {
-			if existing.UUID == entry.UUID {
-				index = idx
-				break
-			}
-
-			if existing.Namespace.ID == entry.Namespace.ID {
-				return errors.New("namespace with specified accessor already exists")
-			}
-
-			if existing.Namespace.Path == entry.Namespace.Path {
-				return errors.New("namespace with specified path already exists")
-			}
+		var existing *NamespaceEntry
+		existing, exists = ns.namespacesByUUID[entry.UUID]
+		if !exists {
+			return errors.New("trying to update a non-existant namespace")
 		}
-	}
 
-	if index != -1 && ns.namespaces[index].Namespace.Path != entry.Namespace.Path {
-		return errors.New("unable to remount namespace at new path")
+		if existing.Namespace.ID != entry.Namespace.ID {
+			return errors.New("accessor ID does not match")
+		}
+
+		if existing.Namespace.Path != entry.Namespace.Path {
+			return errors.New("unable to remount namespace at new path")
+		}
 	}
 
 	if err := ns.writeNamespace(ctx, entry); err != nil {
 		return fmt.Errorf("failed to persist namespace: %w", err)
 	}
 
-	if index == -1 {
-		ns.namespaces = append(ns.namespaces, entry)
-		ns.namespacesByPath[entry.Namespace.Path] = entry
-		ns.namespacesByUUID[entry.UUID] = entry
-		ns.namespacesByAccessor[entry.Namespace.ID] = entry
+	ns.namespacesByPath[entry.Namespace.Path] = entry
+	ns.namespacesByUUID[entry.UUID] = entry
+	ns.namespacesByAccessor[entry.Namespace.ID] = entry
 
-		// Release the lock before creating new entries.
+	// Since the write succeeded, copy back any potentially changed values.
+	nsEntry.UUID = entry.UUID
+	nsEntry.Namespace.ID = entry.Namespace.ID
+	nsEntry.Namespace.Path = entry.Namespace.Path
+
+	if !exists {
+		// unlock before initializeNamespace sice that will re-acqurie the lock
 		ns.lock.Unlock()
 		unlocked = true
 
@@ -326,19 +319,7 @@ func (ns *NamespaceStore) setNamespaceLocked(ctx context.Context, namespace *Nam
 		if err := ns.initializeNamespace(ctx, entry); err != nil {
 			return err
 		}
-	} else {
-		ns.namespaces[index] = entry
-		ns.namespacesByPath[entry.Namespace.Path] = entry
-		ns.namespacesByUUID[entry.UUID] = entry
-		ns.namespacesByAccessor[entry.Namespace.ID] = entry
-
-		// No need to adjust mounts as they should already exist.
 	}
-
-	// Since the write succeeded, copy back any potentially changed values.
-	namespace.UUID = entry.UUID
-	namespace.Namespace.ID = entry.Namespace.ID
-	namespace.Namespace.Path = entry.Namespace.Path
 
 	return nil
 }
@@ -530,7 +511,8 @@ func (ns *NamespaceStore) ModifyNamespaceByPath(ctx context.Context, path string
 		entry = entry.Clone()
 	} else {
 		entry = &NamespaceEntry{Namespace: &namespace.Namespace{
-			Path: path,
+			Path:           path,
+			CustomMetadata: make(map[string]string),
 		}}
 	}
 
@@ -540,6 +522,7 @@ func (ns *NamespaceStore) ModifyNamespaceByPath(ctx context.Context, path string
 		return nil, err
 	}
 
+	// setNamespaceLocked will unlock ns.lock
 	if err := ns.setNamespaceLocked(ctx, entry); err != nil {
 		return nil, err
 	}
@@ -598,8 +581,8 @@ func (ns *NamespaceStore) ListNamespaceEntries(ctx context.Context, includeParen
 	ns.lock.RLock()
 	defer ns.lock.RUnlock()
 
-	entries := make([]*NamespaceEntry, 0, len(ns.namespaces))
-	for _, item := range ns.namespaces {
+	entries := make([]*NamespaceEntry, 0, len(ns.namespacesByUUID))
+	for _, item := range ns.namespacesByUUID {
 		if !includeParent && item.Namespace.ID == parent.ID {
 			continue
 		}
@@ -624,33 +607,22 @@ func (ns *NamespaceStore) DeleteNamespace(ctx context.Context, uuid string) erro
 		return err
 	}
 
+	item, ok := ns.namespacesByUUID[uuid]
+	if !ok {
+		return nil
+	}
+
+	if item.Namespace.ID == namespace.RootNamespaceID {
+		return errors.New("unable to delete root namespace")
+	}
+
 	// Now grab write lock so that we can write to storage.
 	ns.lock.Lock()
 	defer ns.lock.Unlock()
 
-	index := -1
-	for idx, item := range ns.namespaces {
-		if item.UUID == uuid {
-			if item.Namespace.ID == namespace.RootNamespaceID {
-				return errors.New("unable to delete root namespace")
-			}
-
-			index = idx
-			break
-		}
-	}
-
-	if index == -1 {
-		return nil
-	}
-
-	delete(ns.namespacesByPath, ns.namespaces[index].Namespace.Path)
-	delete(ns.namespacesByUUID, ns.namespaces[index].UUID)
-	delete(ns.namespacesByAccessor, ns.namespaces[index].Namespace.ID)
-
-	// We're guaranteed at least one item remaining since the root namespace
-	// should always be present and not be removable.
-	ns.namespaces = append(ns.namespaces[0:index], ns.namespaces[index+1:]...)
+	delete(ns.namespacesByPath, item.Namespace.Path)
+	delete(ns.namespacesByUUID, uuid)
+	delete(ns.namespacesByAccessor, item.Namespace.ID)
 
 	if err := logical.WithTransaction(ctx, ns.storage, func(s logical.Storage) error {
 		storagePath := path.Join(namespaceStoreRoot, uuid)