Resolves #1052 

Benchmarks are run with the following parameters:
```
goos: darwin
goarch: arm64
pkg: github.com/openbao/openbao/vault
cpu: Apple M2 Pro
#namespaces: 1001 (randomly nested)
```

Benchmark before PR:

```
BenchmarkNamespaceStore/GetNamespace-10                           896004              1280 ns/op             152 B/op          4 allocs/op
BenchmarkNamespaceStore/GetNamespaceByAccessor-10                 647536              1798 ns/op             152 B/op          4 allocs/op
BenchmarkNamespaceStore/GetNamespaceByPath-10                    1235167               971.9 ns/op           191 B/op          4 allocs/op
BenchmarkNamespaceStore/ModifyNamespaceByPath-10                     381           3254062 ns/op         2721477 B/op      60715 allocs/op
BenchmarkNamespaceStore/ListAllNamespaces-10                        9542            130236 ns/op          227792 B/op       3322 allocs/op
BenchmarkNamespaceStore/ListNamespaces_non-recursive-10             6057            193162 ns/op          161262 B/op       1979 allocs/op
BenchmarkNamespaceStore/ListNamespaces_recursive-10                97225             13043 ns/op            9701 B/op         25 allocs/op
BenchmarkNamespaceStore/ResolveNamespaceFromRequest-10              9094            130142 ns/op          220279 B/op       3333 allocs/op
BenchmarkNamespaceStore/DeleteNamespace-10                           199           6103086 ns/op         2687740 B/op      59930 allocs/op
```

Benchmark after PR:

```
BenchmarkNamespaceStore/GetNamespace-10                          5931901               205.2 ns/op           152 B/op          4 allocs/op
BenchmarkNamespaceStore/GetNamespaceByAccessor-10                5697356               215.7 ns/op           152 B/op          4 allocs/op
BenchmarkNamespaceStore/GetNamespaceByPath-10                    2128905               564.6 ns/op           302 B/op          6 allocs/op
BenchmarkNamespaceStore/ModifyNamespaceByPath-10                     390           3187719 ns/op         2720228 B/op      60589 allocs/op
BenchmarkNamespaceStore/ListAllNamespaces-10                       10000            107508 ns/op          183808 B/op       2280 allocs/op
BenchmarkNamespaceStore/ListNamespaces_non-recursive-10          1332764               925.1 ns/op           685 B/op         14 allocs/op
BenchmarkNamespaceStore/ListNamespaces_recursive-10               453006              2932 ns/op            2929 B/op         46 allocs/op
BenchmarkNamespaceStore/ResolveNamespaceFromRequest-10           1000000              1131 ns/op             891 B/op         15 allocs/op
BenchmarkNamespaceStore/DeleteNamespace-10                           100          48279762 ns/op          977393 B/op      21774 allocs/op
```

A couple of points to note:

- `ModifyNamespaceByPath` and `DeleteNamespace` are both slowed down due to the inefficient implementation of transactions in the in-memory storage implementation
- The performance difference of `DeleteNamespace` is mainly due to the refactoring adding additional checks (e.g. you can't delete a namespace if it still has child namespaces) and erroring out earlier. Both versions should perform roughly the same due to the transactions.
- Most performance hits left are due to `(*NamespaceEntry).Clone`, which also accounts for most of the memory allocations (except for `ModifyNamespaceByPath` and `DeleteNamespace`, where the allocations are from the transaction)

/cc @cipherboy @driif @satoqz @voigt @wslabosz-reply