@cipherboy 2025-03-17
Just a note, but I don't understand why `BenchmarkNamespaceStore/ModifyNamespaceByPath-10` is so much comically longer than every other operation. It is as expensive as 28 list operations if I read it correctly?