@phyrog 2025-03-17
> Just a note, but I don't understand why `BenchmarkNamespaceStore/ModifyNamespaceByPath-10` is so much comically longer than every other operation. It is as expensive as 28 list operations if I read it correctly? 

That's due to the in-mem transaction implementation. It exports and re-imports the whole storage tree. The benchmarks are run with 1000 namespaces existing. I expect this to perform better with bolt (but haven't figured out how to use that in the benchmark yet).