@cipherboy 2025-03-17
Ahh, yeah, I forgot about that. :D 

Check out:

https://github.com/openbao/openbao/blob/7a1cee6468219fba85a0c0d61281d1061f8a311a/physical/raft/testing.go#L16-L24

and just remember to remove that directory afterwards:

https://github.com/openbao/openbao/blob/7a1cee6468219fba85a0c0d61281d1061f8a311a/physical/crosstest/cross_test.go#L325