@phyrog 2025-03-19
@cipherboy I ran the benchmark locally with raft, but the picture does not look any better. As it turns out (surprise), the filesystem is also slow. So all the read operations are still using the fast in-memory indices, while the write operations are blocked by using the actual physical backend. So I don't expect these to get any better.