@phyrog 2025-03-20
This is ready for review now. I rebased it on the latest namespaces branch and updated the benchmarks above + pushed some additional minor fixes.