@cipherboy 2025-03-24
> @cipherboy I ran the benchmark locally with raft, but the picture does not look any better. As it turns out (surprise), the filesystem is also slow. So all the read operations are still using the fast in-memory indices, while the write operations are blocked by using the actual physical backend. So I don't expect these to get any better.

Yeah, I realized I was also misreading the benchmark stats and thought delete was several orders of magnitude faster than modify, which is why I was expecting it to be faster and thought either inmem was slow or the algorithm for modification was inefficient.