@phyrog 2025-03-25
> @phyrog Hmm, though it looks like there is a new test failure if I'm not mistaken, unless I missed it early and it was fixed in #1157:
> 
> ```
> 
> 2025-03-24T16:21:49.0864830Z ##[group]�[31mFAIL�[0m vault.TestCore_MountEntryView (0.02s)
> 2025-03-24T16:21:49.0865196Z     mount_test.go:1135: 
> 2025-03-24T16:21:49.0865726Z         	Error Trace:	/home/runner/work/openbao/openbao/vault/mount_test.go:1135
> 2025-03-24T16:21:49.0866185Z         	Error:      	Received unexpected error:
> 2025-03-24T16:21:49.0866825Z         	            	trying to update a non-existant namespace
> 2025-03-24T16:21:49.0867205Z         	Test:       	TestCore_MountEntryView
> ```

Should be fixed by #1157, I'll rebase and make sure.